Init demos app
