+++ config/settings/base.py
@@ -85,6 +85,7 @@ LOCAL_APPS = [
     "envergo.urlmappings",
     "envergo.hedges",
     "envergo.petitions",
+    "envergo.demos",
 ]
 # https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
