+++ config/urls_amenagement.py
@@ -6,7 +6,6 @@ from django.utils.translation import gettext_lazy as _
 
 from config.urls import handler500  # noqa
 from envergo.evaluations.views import ShortUrlAdminRedirectView
-from envergo.geodata.views import CatchmentAreaDebug
 from envergo.users.views import NewsletterDoubleOptInConfirmation, NewsletterOptIn
 
 from .urls import urlpatterns as common_urlpatterns
@@ -91,7 +90,7 @@ urlpatterns = [
     path("avis/", include("envergo.evaluations.urls")),
     path(_("moulinette/"), include("envergo.moulinette.urls_amenagement")),
     path(_("geo/"), include("envergo.geodata.urls")),
-    path("demonstrateur-bv/", CatchmentAreaDebug.as_view(), name="2150_debug"),
+    # path("demonstrateur-bv/", CatchmentAreaDebug.as_view(), name="2150_debug"),
     path("newsletter/", NewsletterOptIn.as_view(), name="newsletter_opt_in"),
     path(
         "newsletter/confirmation/",
