+++ envergo/templates/demos/catchment_area.html
@@ -4,7 +4,7 @@
 
 {% block title %}Démonstrateur 2150{% endblock %}
 
-{% block main-classes %}demonstrateur_2150{% endblock %}
+{% block main-classes %}demos{% endblock %}
 
 {% block content %}
   <article>
