+++ envergo/templates/demos/hedge_density.html
@@ -1,0 +1,136 @@
+{% extends 'base.html' %}
+
+{% load static l10n evaluations moulinette %}
+
+{% block title %}Démonstrateur densité de haie{% endblock %}
+
+{% block main-classes %}demos{% endblock %}
+
+{% block content %}
+  <article>
+    <h1>Estimation de densité de haie : démonstrateur</h1>
+    <div class="fr-grid-row fr-grid-row--gutters">
+      <div class="fr-col-7">
+        <form method="get" action="">
+          {{ form.lng.as_hidden }}
+          {{ form.lat.as_hidden }}
+          <label class="step">
+            Double-cliquez ou déplacez le marqueur sur la carte pour ajuster
+            l'emplacement du projet.
+          </label>
+
+          <div class="ratio-4x3 ratio-sm-4x5 fr-my-3w">
+            <div class="ratio-content">
+              <div id="map"></div>
+            </div>
+          </div>
+
+          <button type="submit" class="fr-btn">Valider</button>
+        </form>
+      </div>
+      <div class="fr-col-5">
+
+        {% if result_available %}
+          <h2 class="fr-h4">Résultat du calcul</h2>
+
+          <table>
+            <caption>Résultat du calcul de densité de haie</caption>
+            <thead>
+              <tr>
+                <th></th>
+                <th>Rayon 200 m</th>
+                <th>Rayon 5 km</th>
+              </tr>
+            </thead>
+
+            <tbody>
+              <tr>
+                <th>Linéaire</th>
+                <td></td>
+                <td></td>
+              </tr>
+              <tr>
+                <th>Surface</th>
+                <td></td>
+                <td></td>
+              </tr>
+              <tr>
+                <th>Densité</th>
+                <td></td>
+                <td></td>
+              </tr>
+            </tbody>
+          </table>
+
+        {% else %}
+          {% if form.is_bound %}
+            <h2 class="fr-h4">Département non disponible</h2>
+            <p>Le calcul de densité de haie n'est pas encore disponible dans ce département.</p>
+          {% else %}
+            <h2 class="fr-h4">Essayez notre démonstrateur</h2>
+            <p>Sélectionnez un point sur la carte et lancez la simulation.</p>
+          {% endif %}
+        {% endif %}
+
+        {% if debug %}
+          <h2 class="fr-h4">Données de calcul</h2>
+
+          {% if values %}
+            <table class="fr-table">
+              {% for row in values %}
+                <tr>
+                  {% for value in row %}<td>{{ value }}</td>{% endfor %}
+                </tr>
+              {% endfor %}
+            </table>
+          {% elif flat_values %}
+            {% for value in flat_values %}<span>{{ value }}&nbsp;</span>{% endfor %}
+          {% endif %}
+
+        {% endif %}
+
+        <h2 class="fr-h4">Pour améliorer ce démonstrateur</h2>
+
+        <p>
+          Nous sommes preneurs de tous vos retours, positifs et critiques : contactez-nous à <a href="mailto:contact@envergo.beta.gouv.fr">contact@envergo.beta.gouv.fr</a>
+        </p>
+
+        <p>
+          Merci de joindre des exemples concrets qui fonctionnent ou ne fonctionnent pas, en utilisant le bouton ci-dessous : il vous permet de partager avec nous une simulation.
+        </p>
+
+        <div class="share-btn hide-print fr-my-8w">{% include '_share_url_btn.html' %}</div>
+      </div>
+    </div>
+  </article>
+{% endblock %}
+
+{% block extra_body %}
+  {% include '_share_url_modal.html' with shorten_url=False share_url=share_btn_url %}
+{% endblock %}
+
+{% block extra_css %}
+  <link href="{% static 'leaflet/dist/leaflet.css' %}" rel="stylesheet">
+  <link href="{% static 'leaflet/draw/leaflet.draw.css' %}" rel="stylesheet">
+{% endblock %}
+
+{% block extra_js %}
+  <script defer src="{% static 'leaflet/leaflet.js' %}"></script>
+  <script defer src="{% static 'leaflet/draw/leaflet.draw.js' %}"></script>
+  <script defer src="{% static 'js/libs/leaflet-icon-fix.js' %}"></script>
+  <script defer src="{% static 'js/libs/2150_map.js' %}"></script>
+  <script defer src="{% static 'js/libs/share_url_modal.js' %}"></script>
+
+
+  <script>
+    var DISPLAY_MARKER = {{ display_marker|yesno:'true,false' }};
+    var CENTER_MAP = {lng: {{ center_map.0|unlocalize }}, lat: {{ center_map.1|unlocalize }}};
+    var DEFAULT_ZOOM = {{ default_zoom }};
+    var LAT_FIELD_ID = 'id_lat';
+    var LNG_FIELD_ID = 'id_lng';
+    var POLYGONS = {{ polygons|default:"[]"|safe }};
+    var ENVELOPE = {{ envelope|default:"{}"|safe }};
+    var DEBUG = {{ debug|yesno:'true,false' }};
+  </script>
+
+{% endblock %}
